Moved old test process test back to BaseNeighbourhoodProcessing to make diff smaller
